fx media type

+++ src/core/file_sys/archive_ncch.h
@@ -15,7 +15,7 @@
 // FileSys namespace
 
 namespace Service::FS {
-enum class MediaType : u32;
+enum class MediaType : u8;
 } // namespace Service::FS
 
 namespace FileSys {
